requirements.txt
python-telegram-bot==13.9
youtube_dl==2021.12.17